emg_gesture/requirements.txt
# EMG Gesture Recognition System Requirements

# Deep Learning
torch>=2.0.0
numpy>=1.24.0

# Signal Processing
scipy>=1.10.0

# Serial Communication
pyserial>=3.5

# Configuration
pyyaml>=6.0

# Web Dashboard
fastapi>=0.100.0
uvicorn[standard]>=0.22.0
websockets>=11.0

# Optional: GPU support (uncomment for CUDA)
# torch-cuda>=2.0.0

# Optional: Development tools
# pytest>=7.0.0
# black>=23.0.0
# mypy>=1.0.0
